add node points to proto
